Remove whole tmpdir and outputdir
The default osbuild_store is in tmpdir, files inevitably obtain root
permissions in here as osbuild requires sudo permissions for loopback
mounts. When the python script exits, it tries to recursively remove
this tmpdir and it spews loads of error messages as it cannot remove the
root owned files. This fixes this issue. It is ok to remove this
directory at the end of a build as it is a randomly generated directory
anyway, intended to be recursively deleted at the end of the build.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>