Restructure include handling of snippets
This changes the structure of includes of the various yaml snippets
with the goal of simplifying the files and making manfests more powerful.

The order of evaluation of the various mpp files when handling includes
is a bit weird. First the main yaml has mpp-vars evalutated, and then the
includes are handled in order, evaluating the mpp-vars in the include files.
This means that:
1) The manifest file cannot access any default or arch/target variables like
   "use_aboot", which is problematic for e.g. the qa manifest that want to do
   aboot specific features.
2) The file that sets the defaults has to manually check of each variable
   is set before setting it to the default.

With this change, instead of using the user-supplied manifest as the main
file to osbuild-mpp we instead parse a base manifest which then just
includes all the snippet in order, with the user-supplied manifest
in the right place (after the default and arch/target snippets).

To get the variables in the right place in the evaluation order we
need to split the variable caclculations from the user supplied manfest
into a separate inlude snippet, but we already modify that file in the
code so that is not a large change. Also, we no longer need to inject
the includes in that file now.

To get the order of the "name" variable right we do some custom handling
of it, and we now also pre-initialize it to the manifest basename.

To clear stuff up, some plain data variables like the gpg keys were moved
to a separate include file.

Some stuff from the defaults file (like the uuid computation) was also
moved to computed-vars.ipp