Use Unix Domain Sockets for BlueChi + QM
Relates to: CentOS/automotive/src/automotive-image-builder#16

The default IP for the bluechi-controller of 127.0.0.1 resolves to the
in-qm container loopback, not the 127.0.0.1 of the host. This leads
to the bluechi-agent inside the QM environment not being able to
connect. To solve this, lets use a unix domain socket for BlueChi
as a connection mechanism and mount the uds from the host into the
QM environment.

Signed-off-by: Michael Engel <mengel@redhat.com>